requirements.txt
alembic==1.5.6
appdirs==1.4.4
black==20.8b1
Bootstrap-Flask==1.5.1
click==7.1.2
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.3
login==0.0.6
Mako==1.1.4
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
regex==2020.11.13
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
